api_tests/requirements.txt
Flask
Flask-WTF
psycopg2-binary
bcrypt
gunicorn
requests
python-dotenv
pytest
